A
We selected the terms "Digital Transformation" and "Cultural Industry" due to their prevalent usage in online discourse regarding organizations adopting digital technology. While acknowledging that some sources may not explicitly use these terms when discussing challenges or opportunities, we minimized this issue by utilizing synonymous terms such as "Fail" or "Failure" for identifying challenges, and "Successful" or "Success" for identifying opportunities within digital communities and integrated the failure and the success  with the Performance Metrics, Artificial Intelligence, Internet of Things (IoT), Strategic Policies, Ethical Implications, and User Experience (UX) in the search keywords and the output of this step is two search keywords one for the challenges and the other for the opportunities.
This strategic combination alongside "Digital Transformation" and "Cultural Industry" enhances search precision and ensures the quality of our findings in subsequent stages.